Crate forge_vm

Modules

vm 🔒

Structs

VM

Virtual Machine (VM) designed for 32-bit architecture operations.

Enums

Instruction

Represents the set of all possible instructions for the ForgeVM virtual machine. Each instruction can manipulate registers, perform arithmetic or logical operations, control program flow, or interact with memory.

VmError

Represents the possible errors that can occur during the execution of the VM. The errors are related to memory access, invalid instructions, invalid opcodes, invalid registers, division by zero, stack underflow, stack overflow, and other errors. The Other variant is used to represent any other error that does not fit the other categories. The String field contains a description of the error.