Enum forge_vm::vm::instructions::Instruction

pub enum Instruction<D, A> {

    NOP,
    JMP {
        address: A,
    },
    JMPN {
        address: A,
    },
    JMPP {
        address: A,
    },
    JMPZ {
        address: A,
    },
    CALL {
        address: A,
    },
    RET,
    HLT,
    MOV {
        dest: u8,
        value: D,
    },
    LD {
        dest: u8,
        address: A,
    },
    ST {
        src: u8,
        address: A,
    },
    PUSHREG {
        reg: u8,
    },
    POPREG {
        reg: u8,
    },
    ADD {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    SUB {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    MULT {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    DIV {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    MOD {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    INC {
        reg: u8,
    },
    DEC {
        reg: u8,
    },
    AND {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    OR {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    XOR {
        dest: u8,
        reg1: u8,
        reg2: u8,
    },
    NOT {
        dest: u8,
        reg: u8,
    },
    CMP {
        reg1: u8,
        reg2: u8,
    },
    CLF,

}

Represents the set of all possible instructions for the ForgeVM virtual machine. Each instruction can manipulate registers, perform arithmetic or logical operations, control program flow, or interact with memory.

This enum is used to decode and execute instructions from the bytecode loaded into the VM.

Variants

NOP

No operation. Performs no action and is typically used to consume a cycle without any effect.

JMP

Fields

address: A

Jump to a specified address in the program

Parameters

address: The address to jump to

JMPN

Fields

address: A

Jump if negative flag is set

Parameters

address: The address to jump to if the negative flag is set

JMPP

Fields

address: A

Jump if negative flag is not set

Parameters

address: The address to jump to if the negative flag is not set

JMPZ

Fields

address: A

Jump if zero flag is set

Parameters

address: The address to jump to if the zero flag is set

CALL

Fields

address: A

Call a function at a specified address in the program This operation pushes the current program counter onto the stack and jumps to the specified address.

Parameters

address: The address of the function to call

RET

Return from a function call This operation pops the top of the stack and sets the program counter to the popped value.

HLT

Halt the program execution This operation stops the program execution.

MOV

Fields

dest: u8

value: D

Moves a specified value into the designated dest register.

Parameters

dest: The register number into which the value will be moved.

value: The value to store in the register.

LD

Fields

dest: u8

address: A

Loads a value from the specified address in memory into the dest register. This operation reads the memory at the given address and updates the register with the value found.

Parameters

dest: The destination register where the memory content will be loaded.

address: Memory address from which data is to be read.

ST

Fields

src: u8

address: A

Stores the value from src register into the memory at address. This operation writes the value in the specified register to the specified address in memory.

Parameters

src: The source register whose content is to be stored.

address: Memory address at which the data is to be stored.

PUSHREG

Fields

reg: u8

Push the value from reg register onto the stack. This operation pushes the value from the specified register onto the stack.

Parameters

reg: The register whose value is to be pushed onto the stack.

POPREG

Fields

reg: u8

Pop the value from the top of the stack into the reg register. This operation pops the value from the top of the stack and stores it in the specified register.

Parameters

reg: The register into which the value from the stack is to be popped.

ADD

Fields

dest: u8

reg1: u8

reg2: u8

Add two registers and store the result in a destination register This operation adds the values in two registers and stores the result in the destination register. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be added.

reg2: The second register containing the value to be added.

SUB

Fields

dest: u8

reg1: u8

reg2: u8

Subtract two registers and store the result in a destination register This operation subtracts the value in the second register from the value in the first register and stores the result in the destination register. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be subtracted from.

reg2: The second register containing the value to be subtracted.

MULT

Fields

dest: u8

reg1: u8

reg2: u8

Multiply two registers and store the result in a destination register This operation multiplies the values in two registers and stores the result in the destination register. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be multiplied.

reg2: The second register containing the value to be multiplied.

DIV

Fields

dest: u8

reg1: u8

reg2: u8

Divide two registers and store the result in a destination register This operation divides the value in the first register by the value in the second register and stores the result in the destination register. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the dividend.

reg2: The second register containing the divisor.

MOD

Fields

dest: u8

reg1: u8

reg2: u8

Calculate the remainder of dividing two registers and store the result in a destination register This operation calculates the remainder of dividing the value in the first register by the value in the second register and stores the result in the destination register. If the result is zero, the zero flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the dividend.

reg2: The second register containing the divisor.

INC

Fields

reg: u8

Increment a register This operation increments the value in the specified register by one. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

reg: The register to increment.

DEC

Fields

reg: u8

Decrement a register This operation decrements the value in the specified register by one. If the result is too large to fit in the register, the overflow flag is set. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

reg: The register to decrement.

AND

Fields

dest: u8

reg1: u8

reg2: u8

Logical AND two registers and store the result in a destination register This operation performs a bitwise AND operation on the values in two registers and stores the result in the destination register. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be ANDed.

reg2: The second register containing the value to be ANDed.

OR

Fields

dest: u8

reg1: u8

reg2: u8

Logical OR two registers and store the result in a destination register This operation performs a bitwise OR operation on the values in two registers and stores the result in the destination register. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be ORed.

reg2: The second register containing the value to be ORed.

XOR

Fields

dest: u8

reg1: u8

reg2: u8

Logical XOR two registers and store the result in a destination register This operation performs a bitwise XOR operation on the values in two registers and stores the result in the destination register. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg1: The first register containing the value to be XORed.

reg2: The second register containing the value to be XORed.

NOT

Fields

dest: u8

reg: u8

Logical NOT a register and store the result in a destination register This operation performs a bitwise NOT operation on the value in the specified register and stores the result in the destination register. If the result is zero, the zero flag is set. If the result is negative, the negative flag is set.

Parameters

dest: The destination register where the result will be stored.

reg: The register containing the value to be NOTed.

CMP

Fields

reg1: u8

reg2: u8

Compare two registers This operation compares the values in two registers and sets the zero flag if they are equal.

Parameters

reg1: The first register to compare.

reg2: The second register to compare.

CLF

Clear the flags This operation clears all the flags in the status register.

Implementations

impl<D, A> Instruction<D, A>

pub fn size(&self) -> usize